// packages/database/prisma-generated/zod/inputTypeSchemas/ReviewVoteJoinCountOrderByAggregateInputSchema.ts
import type { Prisma } from '../../client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const ReviewVoteJoinCountOrderByAggregateInputSchema: z.ZodType<Prisma.ReviewVoteJoinCountOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional(),
  updatedAt: z.lazy(() => SortOrderSchema).optional(),
  voteType: z.lazy(() => SortOrderSchema).optional(),
  userId: z.lazy(() => SortOrderSchema).optional(),
  reviewId: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default ReviewVoteJoinCountOrderByAggregateInputSchema;